requirements.txt
Flask==2.3.3
Werkzeug==2.3.7
pyodbc==4.0.39
pandas==2.0.3
openpyxl==3.1.2
python-dotenv==1.0.0
waitress==2.1.2